// src/models/Point.ts
export interface Point {
  id?: string;
  point: string;
  Name: string;
  description: string;
  mainMrd: string;
  imgUrl: string;
  secMrd: string;
  mirror1: string;
  mirror2: string;
  mirror3: string;
  mirror4: string;
  mirror5: string;
  location: string;
  headList: string[];
  neckList: string[];
  armsList: string[];
  bodyList: string[];
  sexsualList: string[];
  legsList: string[];
  internalList: string[];
  backList: string[];
  globalList: string[];
  spiritList: string[];

}